memory management, armor, deep copying entities

+++ headers/entities.h
@@ -8,14 +8,14 @@ typedef struct s_item_group
 {
 	char	*category;
 	t_item	*array;
-	int		count;
+	size_t	count;
 }	t_item_group;
 
 typedef struct s_creature_group
 {
 	char		*category;
 	t_creature	*array;
-	int			count;
+	size_t		count;
 }	t_creature_group;
 
 t_item_group *get_item_group(char *group);
